Moved microbatching into example
